django/requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
et-xmlfile==1.1.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pillow==10.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
